+++ src/Handler/Service/GetByIdHandler.php
@@ -5,7 +5,6 @@ namespace StadGent\Services\OpeningHours\Handler\Service;
 use StadGent\Services\OpeningHours\Handler\HandlerAbstract;
 use StadGent\Services\OpeningHours\Request\Service\GetByIdRequest;
 use StadGent\Services\OpeningHours\Response\ServiceResponse;
-use StadGent\Services\OpeningHours\Response\Validator\ServiceNotFoundValidator;
 use StadGent\Services\OpeningHours\Value\Service;
 use Psr\Http\Message as Psr;
 
@@ -27,7 +26,6 @@ class GetByIdHandler extends HandlerAbstract
     /**
      * @inheritDoc
      *
-     * @throws \StadGent\Services\OpeningHours\Response\Exception\InvalidResponseException
      * @throws \InvalidArgumentException
      */
     public function toResponse(Psr\ResponseInterface $response)